Add files via upload
Five tab-separated CSV-style files and five identical XLSX files, containing the data for these five conferences (and the CHI extended abstracts) since their origins. 

Each file has the same format: 

the first column is full conference proceedings name as in the DL: "Proceedings of the 2018 CHI Conference on Human Factors in Computing Systems"
the second column is the paper title: "Breaking! A Typology of Security and Privacy News and How It's Shared"
the third is the paper's DOI: 10.1145/3173574.3173575
the fourth is the author name, as in the DL: like Sauvik Das
the fifth is the author_url, as in the DL (which is nice for author differentiation): https://dl.acm.org/author_page.cfm?id=81500644254
the sixth is author_position, which says which author this person is on this paper, so it's an integer from 1 to n: 1